[test | refactor] mock-data파일 이름 변경
- mock-dats > client-mock-data로 이름 변경
  - 오타 수정
  - 서버 mock-data파일을 따로 생성할 것임
