+++ src/components/UI/molecule/Dropdown/Dropdown.stories.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { action } from '@storybook/addon-actions';
 import SkillDropdown from '.';
-import { makeSkillMock } from '../../../../__mocks__/mock-dats';
+import { makeSkillMock } from '../../../../__mocks__/client-mock-data';
 
 export default {
   title: 'Molecule/Dropdown',
