+++ src/pages/Signin/Signin.stories.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import MockAdapter from 'axios-mock-adapter';
 import SigninTemplate from './template';
 import AxiosMock from '../../common/AxiosMock';
-import { signinSuccessMockInfo, signinFailMockInfo } from '../../__mocks__/mock-dats';
+import { signinSuccessMockInfo, signinFailMockInfo } from '../../__mocks__/client-mock-data';
 
 export default {
   title: 'Template/Signin',
