+++ src/pages/Signup/Signup.stories.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import MockAdapter from 'axios-mock-adapter';
 import SignupTemplate from './template';
 import AxiosMock from '../../common/AxiosMock';
-import { signupSuccessMockInfo, signupFailMockInfo } from '../../__mocks__/mock-dats';
+import { signupSuccessMockInfo, signupFailMockInfo } from '../../__mocks__/client-mock-data';
 
 export default {
   title: 'Template/Signup',
